refactor: centralize diff stats utilities
