+++ src/agent/output/index.ts
@@ -2,3 +2,4 @@ export * from './OutputHandler';
 export * from './IOutputHandler';
 export { NamedOutputFile } from './types';
 export { getOutputFileName } from '@agent/utils/outputFileUtils';
+export { default as DiffStatsManager } from './DiffStatsManager';
